[Spec 0602] Rename overview-cache to overview-data, fix Team visibility
Rename overview-cache.ts to overview-data.ts. Update all view imports.
Set codev.teamEnabled VS Code context from workspace state so Team
section shows/hides based on actual team configuration.
